urlstore: update a judgement in urlstore
when get one url from crawler queue, if no one exist, it will return
an error causing the invoker terminate.

This commit checks the error message in GetOneNeedCrawlerURL(), if it
failed caused by the error the queue is empty, the function will
return an empty string and no error ("", nil).